✨ Boss Combat Logic, AOE attacks, Boss Animations
Added custom logic for a boss that makes it predict the players movement by getting velocity of the player. Added special attacks for the boss and phase 2 of the boss where it gets stronger at half HP. Added more animations for the boss for variety. Added custom particle for AOE attacks of the boss.